Part5
Added the improved load.js to add records, streamFromDb.js to run the
server that outputs a comma separate values list of users and emails,
and test/testPart5.js that test the server (the server must be running
to do the test, at the moment, I am having troubles starting the server
from the test script). Also, test/testValidEmail.js is used to test the
helpr function isValidEmail(str).